♻️ Split browser updates into a seperate file
