+++ src/cmds.ts
@@ -139,4 +139,11 @@ export const commands: Cmd[] = [
     description: 'Status and files changed for src directory.',
     requestController: async () => (await import('./commands/status')).status,
   },
+  {
+    cmd: 'updates browser',
+    description:
+      'Generate update manifest for the browser binary. This should be run after packaging',
+    requestController: async () =>
+      (await import('./commands/updates/browser')).generateBrowserUpdateFiles,
+  },
 ]
