+++ src/commands/package.ts
@@ -28,31 +28,17 @@ import {
   ReleaseInfo,
   windowsPathToUnix,
 } from '../utils'
+import { generateBrowserUpdateFiles } from './updates/browser'
 
 const machPath = resolve(ENGINE_DIR, 'mach')
 
-/**
- * These are all of the different platforms that aus should deploy to. Note that
- * the names have been simplified and they are all only the ones that are
- * supported by Pulse Browser. If you need something else, open an issue on gh.
- *
- * Based off the code from mozrelease:
- * https://searchfox.org/mozilla-central/source/python/mozrelease/mozrelease/platforms.py
- * https://searchfox.org/mozilla-central/source/taskcluster/gecko_taskgraph/util/partials.py
- */
-const ausPlatformsMap = {
-  linux64: ['Linux_x86_64-gcc3'],
-  macosIntel: [
-    'Darwin_x86_64-gcc3-u-i386-x86_64',
-    'Darwin_x86-gcc3-u-i386-x86_64',
-    'Darwin_x86-gcc3',
-    'Darwin_x86_64-gcc3',
-  ],
-  macosArm: ['Darwin_aarch64-gcc3'],
-  win64: ['WINNT_x86_64-msvc', 'WINNT_x86_64-msvc-x64'],
-}
-
 export const melonPackage = async () => {
+  const brandingKey = dynamicConfig.get('brand') as string
+  const brandingDetails = config.brands[brandingKey]
+
+  const version = brandingDetails.release.displayVersion
+  const channel = brandingKey || 'unofficial'
+
   // The engine directory must have been downloaded for this to be valid
   // TODO: Make this a reusable function that can be used by everything
   if (!existsSync(ENGINE_DIR)) {
@@ -131,112 +117,25 @@ export const melonPackage = async () => {
     }
   }
 
+  const marPath = await createMarFile(version, channel)
+  dynamicConfig.set('marPath', marPath)
+
+  await generateBrowserUpdateFiles()
+
   log.info()
   log.info(`Output written to ${DIST_DIR}`)
-
-  const brandingKey = dynamicConfig.get('brand')
-  const brandingDetails = config.brands[brandingKey]
-
-  const version = brandingDetails.release.displayVersion
-  const channel = brandingKey || 'unofficial'
-
-  const marPath = await createMarFile(version, channel)
-
-  log.info('Creating AUS update files')
-
-  await generateUpdateFile(marPath, version, channel, brandingDetails.release)
 
   log.success('Packaging complected!')
 }
 
-function getReleaseMarName(releaseInfo: ReleaseInfo): string | undefined {
-  if (isAppleSilicon()) {
-    log.askForReport()
-    log.warning('Apple silicon is not yet supported by the distribution script')
-    return
+function getCurrentBrandName(): string {
+  const brand = dynamicConfig.get('brand') as string
+
+  if (brand == 'unofficial') {
+    return 'Nightly'
   }
 
-  switch (process.platform) {
-    case 'win32':
-      return releaseInfo.x86?.windowsMar
-    case 'darwin':
-      return releaseInfo.x86?.macosMar
-    case 'linux':
-      return releaseInfo.x86?.linuxMar
-  }
-}
-
-async function generateUpdateFile(
-  marPath: string,
-  version: string,
-  channel: string,
-  releaseInfo: ReleaseInfo
-) {
-  // We need the sha512 hash of the mar file for the update file. AUS will use
-  // this to ensure that the mar file has not been modified on the distribution
-  // server
-  const marHash = createHash('sha512')
-  marHash.update(await readFile(marPath))
-
-  // We need platform information, primarily for the BuildID, but other stuff
-  // might be helpful later
-  let platformINI = join(OBJ_DIR, 'dist', config.binaryName, 'platform.ini')
-  if (!existsSync(platformINI))
-    platformINI = join(OBJ_DIR, 'dist', 'bin', 'platform.ini')
-
-  const platform = parse((await readFile(platformINI)).toString())
-
-  const releaseMarName = getReleaseMarName(releaseInfo)
-  let completeMarURL = `https://${config.updateHostname || 'localhost:8000'}/${
-    releaseMarName || 'output.mar'
-  }`
-
-  // The user is using github to distribute release binaries for this version.
-  if (releaseInfo.github) {
-    completeMarURL = `https://github.com/${releaseInfo.github.repo}/releases/download/${version}/${releaseMarName}`
-    log.info(`Using '${completeMarURL}' as the distribution url`)
-  } else {
-    log.warning(
-      `No release information found! Default release location will be "${completeMarURL}"`
-    )
-  }
-
-  const updateObject = {
-    updates: {
-      update: {
-        // TODO: Correct update type from semvar, store the old version somewhere
-        '@type': 'minor',
-        '@displayVersion': version,
-        '@appVersion': version,
-        '@platformVersion': config.version.version,
-        '@buildID': platform.Build.BuildID,
-
-        patch: {
-          // TODO: Partial patches might be nice for download speed
-          '@type': 'complete',
-          '@URL': completeMarURL,
-          '@hashFunction': 'sha512',
-          '@hashValue': marHash.digest('hex'),
-          '@size': (await stat(marPath)).size,
-        },
-      },
-    },
-  }
-
-  for (const target of getArchFolders()) {
-    const xmlPath = join(
-      DIST_DIR,
-      'update',
-      'browser',
-      target,
-      channel,
-      'update.xml'
-    )
-    const document = create(updateObject)
-
-    ensureEmpty(dirname(xmlPath))
-    await writeFile(xmlPath, document.end({ prettyPrint: true }))
-  }
+  return config.brands[brand].brandFullName
 }
 
 async function createMarFile(version: string, channel: string) {
@@ -280,33 +179,4 @@ async function createMarFile(version: string, channel: string) {
     },
   })
   return marPath
-}
-
-function getArchFolders(): string[] {
-  if (process.platform == 'win32') {
-    return ausPlatformsMap.win64
-  }
-
-  if (process.platform == 'linux') {
-    return ausPlatformsMap.linux64
-  }
-
-  // Everything else will have to be darwin of some kind. So, for future possible
-  // Apple silicon support, we should chose between the two wisely
-  // TODO: This is a hack, fix it
-  if (isAppleSilicon()) {
-    return ausPlatformsMap.macosArm
-  }
-
-  return ausPlatformsMap.macosIntel
-}
-
-function getCurrentBrandName(): string {
-  const brand = dynamicConfig.get('brand')
-
-  if (brand == 'unofficial') {
-    return 'Nightly'
-  }
-
-  return config.brands[brand].brandFullName
 }
